requirements.txt
black
flake8
h5py
isort
matplotlib
numpy
pandas
Pillow
pre-commit
scipy
sklearn
